Feature: Add strict mode
Enable strict mode to immediately throw exceptions that occur during the
requiring process. This makes it easy to find places where explicit
requires are needed, to ensure direct dependencies are loaded.

By default, require_all will catch (and swallow) errors such as NameError
that occur when a file is required, but keep track of the files that didn't
load successfully and retry them later.

This approach, while convenient, can be dangerous and lead to subtle and/or
confusing issues, as it subverts the guarantee of require that a given file
will only be loaded once, a guarantee that code might (fairly) depend on.

For example, if an error such as a NameError occurs during the require of a
file (because an as yet unloaded constant is referenced) that class ends up
being partially defined - the constant for that class exists, but only the
code up to the point of the error will have run. When that file is required
again (ie when retried) the same code that executed previously will run
_again_ too. This will cause any constants that were already defined in the
class to be redefined (issuing warnings from Ruby) but more seriously, if
another class uses that class in its partially defined state, some code that
it depends on (eg methods being defined) may not exist yet, even though the
constant for the class does. There is no easy way to fix this, as the only
way of seeing what the result of requiring a file would be is to require it!